Update Welcome page component
