+++ frontend/src/pages/welcome/index.ts
@@ -1,0 +1,1 @@
+export { Welcome as default } from "./welcome.component";
